Use an external Docker bridge network
Fix #5

This is needed to support systems which
use an OpenVPN client and maybe other
scenarios.

The external Docker network is called:
  `netpen-docker-compose-net`

It is currently hard-coded to a certain
subnet, but that could be improved later:
  `10.0.12.1/24`

Add new dev/Makefile target to force rebuild
Docker images:
  `build-dev`

Add documentation to README.md for the
`build-dev` make target.