web: examples/shares: display R/O on settings change too
"Pen" component only handles the items list. In order for settings
changes to trigger this warning, need to move this logic to PenEditor.

In addition, there's no need to pass another 'onUpdate' function as
not passing an 'update' function means this is a read only pen and
therefore should have this warning. This is a better place for this
logic.