web: models: validate objects when converting to dictionary
Error scenario when deleting a namespace containing an enslaved device,
e.g. in the following configuration:

items:
  - netns:
      name: a
  - netns:
      name: b
  - veth:
      name: ab
      dev1:
        netns: netns.a
      dev2:
        netns: netns.b
  - bridge:
      name: br
      netns: netns.b
      ports:
        - veth.ab.dev2

Deleting netns 'a' would result in a crash, since the bridge ports would
still contain `veth.ab.dev2` since the validity of `veth.ab` wasn't checked.

Same scenario would happen if `veth.ab` was enslaved to a vrf.

Fixed by calling `toDict()` when getDictIdbyId() is called. Ideally would
prefer an `isValid()` method on each object but this is a step in the right
direction.